Fixes the issue mistakenly attributed to uglify.
Turns out the minification issue was not a Lineman configuration problem

After lots of experimentation, we isolated the issue to ngmin's 
inability to recognize that FooController was an angular unit.
Using the more formal declaration and attaching it to the app made
running the app out of dist directory with a static file server work
just fine!